+++ src/main/java/cc/cassian/raspberry/compat/EmiCompat.java
@@ -6,7 +6,6 @@ import cofh.ensorcellation.init.EnsorcEnchantments;
 import com.brokenkeyboard.usefulspyglass.UsefulSpyglass;
 import com.simibubi.create.AllItems;
 import com.teamabnormals.allurement.core.registry.AllurementEnchantments;
-import de.cadentem.additional_enchantments.data.AEEffectTags;
 import de.cadentem.additional_enchantments.registry.AEEnchantments;
 import dev.emi.emi.EmiUtil;
 import dev.emi.emi.api.EmiEntrypoint;
@@ -21,8 +20,8 @@ import dev.emi.emi.api.stack.EmiStack;
 import dev.emi.emi.api.widget.WidgetHolder;
 import net.mehvahdjukaar.supplementaries.reg.ModRegistry;
 import net.minecraft.core.Registry;
+import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -41,6 +40,8 @@ import java.util.List;
 
 @EmiEntrypoint
 public class EmiCompat implements EmiPlugin {
+    public static EmiRecipeCategory ANVIL = new EmiRecipeCategory(RaspberryMod.locate("anvil"), EmiStack.of(Items.ANVIL));
+
     public static class EmiSmithingRecipe implements EmiRecipe {
         private final EmiIngredient input1;
         private final EmiStack input2;
@@ -90,6 +91,18 @@ public class EmiCompat implements EmiPlugin {
             widgets.addGeneratedSlot((r) -> this.input1, this.uniq, 0, 0);
             widgets.addSlot(this.input2, 49, 0);
             widgets.addGeneratedSlot((r) -> this.output, this.uniq, 107, 0).recipeContext(this);
+        }
+    }
+
+    public static class EmiAnvilRecipe extends EmiSmithingRecipe {
+
+        public EmiAnvilRecipe(EmiIngredient input1, EmiStack input2, EmiStack output, ResourceLocation id) {
+            super(input1, input2, output, id);
+        }
+
+        @Override
+        public EmiRecipeCategory getCategory() {
+            return EmiCompat.ANVIL;
         }
     }
 
@@ -148,6 +161,54 @@ public class EmiCompat implements EmiPlugin {
         if (ModCompat.CREATE && ModCompat.ENSORCELLATION && ModCompat.SUPPLEMENTARIES && ModCompat.ALLUREMENT) {
             addEnchantments(emiRegistry);
         }
+        if (ModCompat.QUARK) {
+            emiRegistry.addWorkstation(EmiCompat.ANVIL, EmiStack.of(Items.ANVIL));
+            emiRegistry.addWorkstation(EmiCompat.ANVIL, EmiStack.of(Items.CHIPPED_ANVIL));
+            emiRegistry.addWorkstation(EmiCompat.ANVIL, EmiStack.of(Items.DAMAGED_ANVIL));
+            addRunes(emiRegistry);
+        }
+    }
+
+    public final static String[] VANILLA_DYES = {
+            "white",
+            "light_gray",
+            "gray",
+            "black",
+            "brown",
+            "red",
+            "orange",
+            "yellow",
+            "lime",
+            "green",
+            "cyan",
+            "light_blue",
+            "blue",
+            "purple",
+            "magenta",
+            "pink",
+    };
+
+    private void addRunes(EmiRegistry emiRegistry) {
+        for (String dye : VANILLA_DYES) {
+            var enchantedGear = new ItemStack(Items.DIAMOND_CHESTPLATE);
+            enchantedGear.enchant(Enchantments.PROJECTILE_PROTECTION, 1);
+            var runedGear = enchantedGear.copy();
+            var compound = runedGear.getOrCreateTag();
+            var quark = new CompoundTag();
+            var rune = Registry.ITEM.get(new ResourceLocation("quark", "%s_rune".formatted(dye)));
+            quark.putString("id", "quark:%s_rune".formatted(dye));
+            quark.putByte("Count", Byte.parseByte("64"));
+            compound.put("quark:RuneColor", quark);
+            compound.putByte("quark:RuneAttached", Byte.parseByte("1"));
+            runedGear.setTag(compound);
+            emiRegistry.addRecipe(new EmiAnvilRecipe(
+                    EmiStack.of(enchantedGear),
+                    EmiStack.of(new ItemStack(rune)),
+                    EmiStack.of(runedGear),
+                    RaspberryMod.locate("/runes/"+dye)
+            ));
+        }
+
     }
 
     public void addEnchantments(EmiRegistry emiRegistry) {